+++ frontend/src/toolbar/actions/SelectorEditingModal.tsx
