+++ frontend/src/toolbar/elements/Elements.tsx
@@ -4,12 +4,12 @@ import { useActions, useValues } from 'kea'
 import { compactNumber } from 'lib/utils'
 import React from 'react'
 
+import { ElementInfoWindow } from '~/toolbar/elements/ElementInfoWindow'
 import { elementsLogic } from '~/toolbar/elements/elementsLogic'
 import { FocusRect } from '~/toolbar/elements/FocusRect'
 import { HeatmapElement } from '~/toolbar/elements/HeatmapElement'
 import { HeatmapLabel } from '~/toolbar/elements/HeatmapLabel'
 import { heatmapLogic } from '~/toolbar/elements/heatmapLogic'
-import { InfoWindow } from '~/toolbar/elements/InfoWindow'
 import { getBoxColors, getHeatMapHue } from '~/toolbar/utils'
 
 export function Elements(): JSX.Element {
@@ -37,7 +37,7 @@ export function Elements(): JSX.Element {
                     zIndex: 2147483021,
                 }}
             >
-                <InfoWindow />
+                <ElementInfoWindow />
             </div>
             <div
                 id="posthog-toolbar-elements"
